updated to remove old files

+++ DesignCode1/SimpleFactory/OpenGLRenderer.h
@@ -9,8 +9,8 @@ public:
 
   bool loadScene(const std::string &_filename) {return true;}
   void setViewportSize(int _w, int _h) {;}
-  void setCameraPos(double _x, double _y, double _z) {;}
-  void setLookAt(double _x, double _y, double _z) {;}
+  void setCameraPos(float _x, float _y, float _z) {;}
+  void setLookAt(float _x, float _y, float _z) {;}
   void render() {std::cout<<"OpenGL Render\n";}
   ~OpenGLRenderer(){std::cout<<"OpenGL dtor called\n";}
 };
